Changed HP code & Added UI transitions bwterrn playing and placing mode
Also made a new folder for earlier versions of scenes

+++ Assets/Script/UI_HealthDisplay.cs
@@ -13,7 +13,8 @@ public class UI_HealthDisplay : MonoBehaviour
 
 		if (hitPointToTrack && hitPointImages.Length > 0)
 		{
-			int barCount = Mathf.FloorToInt((hitPointToTrack.hitPoint / hitPointToTrack.hitPointMax) * hitPointImages.Length + 0.1f);
+			int barCount = Mathf.FloorToInt(((float)hitPointToTrack.hitPoint / (float)hitPointToTrack.hitPointMax) * hitPointImages.Length + 0.1f);
+			//Debug.Log("barCount: " + barCount + ", (f) (h/Max)" + ((float) hitPointToTrack.hitPoint / (float) hitPointToTrack.hitPointMax) + ", whole thing without + .1: " + ((float)(hitPointToTrack.hitPoint / hitPointToTrack.hitPointMax) * hitPointImages.Length));
 			for (int i = 0; i < hitPointImages.Length; i++)
 			{
 				hitPointImages[i].gameObject.SetActive(i < barCount);
